adds a test for updating nested values
It's currently possible to send an update like:

- (json Update version): {"a.b": 7, "a.c": "ServerTimestamp"}
- (fieldpath UpdatePath version):
    path: "a.b"
    value: 7
    path: "a.c"
    value: "ServerTimestamp"

This is saying: update key "b" within key "a" with value 7, and
vice-versa for key "c" within key "a" with value "ServerTimestamp".

This particular conformance test ensures that the update that gets sent
- especially in the json case - is tightly scoped to a.b, and that a
field mask is not added for "a". Currently, a bug in python does the
latter - see
https://github.com/googleapis/google-cloud-python/issues/7215.
That is, it sends an update like:

path: a.b
value: 7
path: a.c
value: ServerTimestamp

Instead of:

path: a
value: {"b": 7}

Or:

path: a
value: {"b": "ServerTimestamp"}

They look the same, but the former _only_ updates a.b to 7, whereas the
latter replaces the entirety of a with {"b": 7}.